+++ packages/apollos-data-schema/src/index.js
@@ -381,6 +381,7 @@ export const contentItemSchema = gql`
   interface ContentItem {
     id: ID!
     title(hyphenated: Boolean): String
+    publishDate: String
     coverImage: ImageMedia
     images: [ImageMedia]
     videos: [VideoMedia]
@@ -402,6 +403,7 @@ export const contentItemSchema = gql`
   type UniversalContentItem implements ContentItem & Node {
     id: ID!
     title(hyphenated: Boolean): String
+    publishDate: String
     coverImage: ImageMedia
     images: [ImageMedia]
     videos: [VideoMedia]
@@ -423,6 +425,7 @@ export const contentItemSchema = gql`
   type DevotionalContentItem implements ContentItem & Node {
     id: ID!
     title(hyphenated: Boolean): String
+    publishDate: String
     coverImage: ImageMedia
     images: [ImageMedia]
     videos: [VideoMedia]
@@ -446,6 +449,7 @@ export const contentItemSchema = gql`
   type MediaContentItem implements ContentItem & Node {
     id: ID!
     title(hyphenated: Boolean): String
+    publishDate: String
     coverImage: ImageMedia
     images: [ImageMedia]
     videos: [VideoMedia]
@@ -469,6 +473,7 @@ export const contentItemSchema = gql`
   type ContentSeriesContentItem implements ContentItem & Node {
     id: ID!
     title(hyphenated: Boolean): String
+    publishDate: String
     coverImage: ImageMedia
     images: [ImageMedia]
     videos: [VideoMedia]
@@ -493,6 +498,7 @@ export const contentItemSchema = gql`
   type WeekendContentItem implements ContentItem & Node {
     id: ID!
     title(hyphenated: Boolean): String
+    publishDate: String
     coverImage: ImageMedia
     images: [ImageMedia]
     videos: [VideoMedia]
@@ -552,6 +558,7 @@ export const contentChannelSchema = gql`
 
   extend type Query {
     contentChannels: [ContentChannel]
+      @deprecated(reason: "No longer supported.")
   }
 `;
 
@@ -805,6 +812,7 @@ export const featuresSchema = gql`
     READ_CONTENT
     READ_EVENT
     OPEN_URL
+    OPEN_AUTHENTICATED_URL
     OPEN_NODE
     OPEN_CHANNEL
   }
@@ -968,6 +976,48 @@ export const featuresSchema = gql`
   }
 `;
 
+export const commentSchema = gql`
+  extend type Mutation {
+    addComment(
+      parentId: ID!
+      text: String!
+      visibility: CommentVisibility
+    ): Comment
+
+    flagComment(commentId: ID!): Comment
+  }
+
+  type CommentListFeature implements Feature & Node {
+    id: ID!
+    order: Int
+
+    comments: [Comment] @cacheControl(maxAge: 0)
+  }
+
+  type AddCommentFeature implements Feature & Node {
+    id: ID!
+    order: Int
+    relatedNode: Node!
+
+    addPrompt: String
+    initialPrompt: String
+  }
+
+  enum CommentVisibility {
+    PUBLIC
+    PRIVATE
+    FOLLOWERS
+  }
+
+  type Comment implements Node {
+    id: ID!
+
+    person: Person
+    text: String
+    visibility: CommentVisibility
+  }
+`;
+
 export const eventSchema = gql`
   type Event implements Node {
     id: ID!
@@ -1002,12 +1052,24 @@ export const prayerSchema = gql`
     prayers: [PrayerRequest]
   }
 
+  type VerticalPrayerListFeature implements Feature & Node {
+    id: ID!
+    order: Int
+    title: String
+    subtitle: String
+    prayers: [PrayerRequest]
+  }
+
   extend type Mutation {
     addPrayer(text: String!, isAnonymous: Boolean): PrayerRequest
   }
 
   extend enum InteractionAction {
     PRAY
+  }
+
+  extend type Person {
+    prayers: [PrayerRequest]
   }
 `;