+++ packages/apollos-ui-connected/src/FeaturesFeedConnected/featuresFeedComponentMapper.js
@@ -1,15 +1,19 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
+import { named } from '@apollosproject/ui-kit';
 import ActionListFeatureConnected from '../ActionListFeatureConnected';
 import HorizontalCardListFeatureConnected from '../HorizontalCardListFeatureConnected';
 import VerticalCardListFeatureConnected from '../VerticalCardListFeatureConnected';
 import HeroListFeatureConnected from '../HeroListFeatureConnected';
 import PrayerListFeatureConnected from '../PrayerListFeatureConnected';
+import VerticalPrayerListFeatureConnected from '../VerticalPrayerListFeatureConnected';
 import ActionBarFeatureConnected from '../ActionBarFeatureConnected';
 import ScriptureFeature from '../features/ScriptureFeature';
 import TextFeature from '../features/TextFeature';
 import WebviewFeature from '../features/WebviewFeature';
+import CommentListFeatureConnected from '../CommentListFeatureConnected';
+import AddCommentFeatureConnected from '../AddCommentFeatureConnected';
 
 const MAPPINGS = {
   ActionListFeature: ActionListFeatureConnected,
@@ -21,6 +25,9 @@ const MAPPINGS = {
   // like Prayer, don't all get stuck here, forcing all church apps to have
   // them all installed?
   PrayerListFeature: PrayerListFeatureConnected,
+  VerticalPrayerListFeature: VerticalPrayerListFeatureConnected,
+  CommentListFeature: CommentListFeatureConnected,
+  AddCommentFeature: AddCommentFeatureConnected,
   ScriptureFeature,
   TextFeature,
   WebviewFeature,
@@ -63,4 +70,6 @@ featuresFeedComponentMapper.propTypes = {
   refetchRef: PropTypes.func.isRequired,
 };
 
-export default featuresFeedComponentMapper;
+export default named('ui-connected.FeaturesFeed.FeatureFeedComponentMapper')(
+  featuresFeedComponentMapper
+);