+++ packages/apollos-ui-connected/src/NodeFeaturesConnected/NodeFeaturesConnected.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Query } from '@apollo/client/react/components';
-import { ErrorCard } from '@apollosproject/ui-kit';
+import { ErrorCard, named } from '@apollosproject/ui-kit';
 import { get } from 'lodash';
 
 import FeaturesFeedConnected from '../FeaturesFeedConnected';
@@ -17,7 +17,6 @@ const NodeFeaturesConnected = ({ Component, nodeId, ...props }) => {
       variables={{ nodeId }}
     >
       {({ data: { node } = {}, loading, error }) => {
-        console.log(error);
         if (error) return <ErrorCard error={error} />;
         // TODO: set an optimistic response for this query to return a visually appeallying empty query so we can enable loading states
         if (loading) return null;
@@ -41,4 +40,6 @@ NodeFeaturesConnected.defaultProps = {
   Component: FeaturesFeedConnected,
 };
 
-export default NodeFeaturesConnected;
+export default named('ui-connected.NodeFeaturesConnected')(
+  NodeFeaturesConnected
+);