+++ packages/apollos-ui-connected/src/NodeSingleConnected/index.js
@@ -1,9 +1,14 @@
 import React from 'react';
-import { Animated } from 'react-native';
+import { Animated, View } from 'react-native';
 import PropTypes from 'prop-types';
 import { Query } from '@apollo/client/react/components';
 
-import { styled, BackgroundView, StretchyView } from '@apollosproject/ui-kit';
+import {
+  styled,
+  BackgroundView,
+  StretchyView,
+  named,
+} from '@apollosproject/ui-kit';
 import { ApollosPlayerContainer } from '@apollosproject/ui-media-player';
 
 import ContentNodeConnected from '../ContentNodeConnected';
@@ -11,6 +16,7 @@ import ContentParentFeedConnected from '../ContentParentFeedConnected';
 import ContentChildFeedConnected from '../ContentChildFeedConnected';
 import UpNextButtonConnected from '../UpNextButtonConnected';
 import NodeFeaturesConnected from '../NodeFeaturesConnected';
+import ScriptureNodeConnected from '../ScriptureNodeConnected';
 
 import GET_MEDIA from './getMedia';
 
@@ -18,17 +24,18 @@ const Noop = () => null;
 
 const FlexedScrollView = styled({ flex: 1 })(Animated.ScrollView);
 
-const NodeSingleInner = ({ nodeId, ImageWrapperComponent }) => (
-  <>
+const NodeSingleInner = ({ nodeId, ImageWrapperComponent, ...props }) => (
+  <View {...props}>
     <ContentNodeConnected
       ImageWrapperComponent={ImageWrapperComponent}
       nodeId={nodeId}
     />
+    <ScriptureNodeConnected nodeId={nodeId} />
     <NodeFeaturesConnected nodeId={nodeId} />
     <UpNextButtonConnected nodeId={nodeId} />
     <ContentParentFeedConnected nodeId={nodeId} />
     <ContentChildFeedConnected nodeId={nodeId} />
-  </>
+  </View>
 );
 
 NodeSingleInner.propTypes = {
@@ -36,42 +43,74 @@ NodeSingleInner.propTypes = {
   ImageWrapperComponent: PropTypes.any, // eslint-disable-line
 };
 
-const NodeSingleConnected = ({ nodeId }) => (
-  <BackgroundView>
-    <StretchyView>
-      {({ Stretchy, ...scrollViewProps }) => (
-        <FlexedScrollView {...scrollViewProps}>
-          <NodeSingleInner nodeId={nodeId} ImageWrapperComponent={Stretchy} />
-        </FlexedScrollView>
-      )}
-    </StretchyView>
-  </BackgroundView>
+const NodeSingleConnected = ({ nodeId, children, Component, ...props }) => (
+  <>
+    <BackgroundView>
+      <StretchyView>
+        {({ Stretchy, ...scrollViewProps }) => (
+          <FlexedScrollView {...scrollViewProps}>
+            <Component
+              nodeId={nodeId}
+              ImageWrapperComponent={Stretchy}
+              {...props}
+            />
+          </FlexedScrollView>
+        )}
+      </StretchyView>
+    </BackgroundView>
+    {children}
+  </>
 );
 
 NodeSingleConnected.propTypes = {
   nodeId: PropTypes.string,
+  children: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+  Component: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
 };
 
-const NodeSingleConnectedWithMedia = ({ nodeId }) => (
+NodeSingleConnected.defaultProps = {
+  Component: NodeSingleInner,
+};
+
+const NodeSingleConnectedWithMedia = ({
+  nodeId,
+  children,
+  Component,
+  ...props
+}) => (
   <Query
     query={GET_MEDIA}
     variables={{ nodeId }}
     fetchPolicy={'cache-and-network'}
   >
     {({ data = {} }) => {
-      if (!data?.node?.videos?.length)
-        return <NodeSingleConnected nodeId={nodeId} />;
+      const hasMedia =
+        data?.node?.videos?.length &&
+        data.node.videos.some(({ sources }) => sources.length);
 
+      if (!hasMedia)
+        return (
+          <NodeSingleConnected nodeId={nodeId} {...props}>
+            {children}
+          </NodeSingleConnected>
+        );
       return (
-        <ApollosPlayerContainer
-          source={data.node?.videos[0]?.sources[0]}
-          coverImage={data.node?.coverImage?.sources}
-          presentationProps={{
-            title: data.node.title,
-          }}
-        >
-          <NodeSingleInner nodeId={nodeId} ImageWrapperComponent={Noop} />
-        </ApollosPlayerContainer>
+        <BackgroundView>
+          <ApollosPlayerContainer
+            source={data.node?.videos[0]?.sources[0]}
+            coverImage={data.node?.coverImage?.sources}
+            presentationProps={{
+              title: data.node.title,
+            }}
+          >
+            <Component
+              nodeId={nodeId}
+              ImageWrapperComponent={Noop}
+              {...props}
+            />
+          </ApollosPlayerContainer>
+          {children}
+        </BackgroundView>
       );
     }}
   </Query>
@@ -79,6 +118,14 @@ const NodeSingleConnectedWithMedia = ({ nodeId }) => (
 
 NodeSingleConnectedWithMedia.propTypes = {
   nodeId: PropTypes.string,
+  Component: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+  children: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
 };
 
-export default NodeSingleConnectedWithMedia;
+NodeSingleConnectedWithMedia.defaultProps = {
+  Component: NodeSingleInner,
+};
+
+export default named('ui-connected.NodeSingleConnected')(
+  NodeSingleConnectedWithMedia
+);