+++ packages/apollos-ui-htmlview/src/HTMLView/defaultRenderer.js
@@ -23,6 +23,10 @@ const LINE_BREAK = '\n';
 const TEXT_TYPES_THAT_SHOULD_WRAP = [Text, BodyText, ButtonLink];
 const ILLEGAL_TEXT_CHILDREN_TYPES = [ConnectedImage, View];
 
+// Detects if the image urls are relative.
+// Example /images/foo.png vs //rock.com/images/foo.png
+export const isLocalImg = (src) => src.startsWith('/') && !src.startsWith('//');
+
 export const stripIllegalMarkup = (children) =>
   Children.toArray(children).filter(
     (child) => !ILLEGAL_TEXT_CHILDREN_TYPES.includes(child.type)
@@ -61,15 +65,30 @@ export const wrapTextChildren = ({
 };
 
 const defaultRenderer = (node, { children }, handlePressAnchor) => {
-  if (node.type === 'text' && node.data && node.data.trim()) {
+  const blockElements = ['p', 'div', 'blockquote'];
+
+  if (node.type === 'text' && node.data) {
     const text = decodeHTML(node.data);
-    if (!node.parent) {
+    const blankText = !text.trim();
+
+    if (
+      blankText &&
+      ((node.next && !blockElements.includes(node.next.name)) ||
+        (node.prev && !blockElements.includes(node.prev.name)))
+    ) {
+      // If the text is blank, but it's beside other text, include it.
+      // Think: two spans with a space between them.
+      return <Text>{text}</Text>;
+    }
+
+    if (!node.parent && !blankText) {
+      // If the text top level, show it inside a body text element.
       return <BodyText>{text}</BodyText>;
     }
-    return <Text>{text}</Text>;
+    if (!blankText) {
+      return <Text>{text}</Text>;
+    }
   }
-
-  const blockElements = ['p', 'div', 'blockquote'];
 
   switch (node.name) {
     case 'div':
@@ -91,7 +110,7 @@ const defaultRenderer = (node, { children }, handlePressAnchor) => {
     case 'blockquote':
       return (
         <BlockQuote>
-          {wrapTextChildren({ children, Component: Text, strip: false })}
+          {wrapTextChildren({ children, Component: BodyText, strip: false })}
         </BlockQuote>
       );
     case 'h1':
@@ -148,6 +167,8 @@ const defaultRenderer = (node, { children }, handlePressAnchor) => {
       const source = {
         url: node.attribs.src,
       };
+
+      if (source.url && isLocalImg(source.url)) return null;
 
       const imgStyles = {
         resizeMode: 'contain',