+++ packages/apollos-data-schema/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0](https://github.com/apollosproject/apollos-apps/compare/v2.0.0-beta.4...v2.0.0) (2021-02-03)
+
+**Note:** Version bump only for package @apollosproject/data-schema
+
+
+
+
+
+# [2.0.0-beta.2](https://github.com/apollosproject/apollos-apps/compare/v2.0.0-beta.1...v2.0.0-beta.2) (2021-01-27)
+
+
+### Features
+
+* Adds User Prayer List ([#1690](https://github.com/apollosproject/apollos-apps/issues/1690)) ([6830f79](https://github.com/apollosproject/apollos-apps/commit/6830f79892552d04eb3e4ce79d19c7f2a802e87f))
+* Comment visibility and flagging ([#1693](https://github.com/apollosproject/apollos-apps/issues/1693)) ([61e0386](https://github.com/apollosproject/apollos-apps/commit/61e03861bb8b09684be4b6b0b89fbf739062988d))
+
+
+
+
+
+# [2.0.0-beta.1](https://github.com/apollosproject/apollos-apps/compare/v2.0.0-beta.0...v2.0.0-beta.1) (2021-01-22)
+
+
+### Features
+
+* Comments ([#1685](https://github.com/apollosproject/apollos-apps/issues/1685)) ([9e12b69](https://github.com/apollosproject/apollos-apps/commit/9e12b69d167f12fa42df2937cf636191b9653cb4))
+
+
+
+
+
+# [1.8.0-beta.0](https://github.com/apollosproject/apollos-apps/compare/v1.7.0...v1.8.0-beta.0) (2021-01-06)
+
+
+### Features
+
+* add date to content items ([#1614](https://github.com/apollosproject/apollos-apps/issues/1614)) ([5345a55](https://github.com/apollosproject/apollos-apps/commit/5345a55))
+* Allows Action URLs to be Authenticated ([#1669](https://github.com/apollosproject/apollos-apps/issues/1669)) ([d223513](https://github.com/apollosproject/apollos-apps/commit/d223513))
+
+
+
+
+
 # [1.7.0](https://github.com/apollosproject/apollos-apps/compare/v1.7.0-beta.3...v1.7.0) (2020-11-18)
 
 **Note:** Version bump only for package @apollosproject/data-schema